fix: verbosity of configuration cli

+++ tergite_autocalibration/config/package.py
@@ -99,6 +99,7 @@ class ConfigurationPackage:
                 # Check whether the files exist in the meta configuration file
                 if file_key_ in meta_config["files"].keys():
                     # Set the filepaths in the attribute
+                    logging.info(f"Loading {file_key_}: {meta_config['files'][file_key_]}")
                     return_obj.config_files[file_key_] = os.path.join(
                         config_path_prefix, meta_config["files"][file_key_]
                     )
@@ -133,6 +134,7 @@ class ConfigurationPackage:
 
         # Unzip the archive in the way it works on macOS such that it creates a folder with the same name
         meta_config_folder_path = os.path.splitext(meta_config_zip_path)[0]
+        logging.info(f"Unpacking configuration package from {meta_config_zip_path} to {meta_config_folder_path}")
         shutil.unpack_archive(meta_config_zip_path, meta_config_folder_path)
 
         # Load the configuration from the .toml file inside
@@ -193,6 +195,11 @@ class ConfigurationPackage:
             # Ensure that the parent directory exists
             os.makedirs(os.path.split(new_file_path)[0], exist_ok=True)
             # Copy the file to the new directory
+            logging.info(
+                f"Copying: \n"
+                f"{file_path} \n"
+                f"-> {new_file_path}"
+            )
             shutil.copy(file_path, new_file_path)
 
         # Iterate over all misc folders and copy them
@@ -200,9 +207,19 @@ class ConfigurationPackage:
             # Get the relative path from the current meta configuration to the misc folder
             rel_path = os.path.relpath(misc_filepath, old_path_to_meta)
             # Then create the new path to the location
-            new_filepath = os.path.join(abs_path_to, rel_path)
+            new_filepath = os.path.join(str(abs_path_to), str(rel_path))
             # Copy the whole folder to the new location
-            shutil.copytree(misc_filepath, new_filepath, dirs_exist_ok=True)
+            if os.path.exists(new_filepath):
+                logging.warning(f"Location {new_filepath} already exists for the configuration package to copy. "
+                                f"The files will not be copied to avoid conflicts. "
+                                f"Please rename {misc_filepath} if you want to have the folder copied.")
+            else:
+                logging.info(
+                    f"Copying: \n"
+                    f"{misc_filepath} \n"
+                    f"-> {new_filepath}"
+                )
+                shutil.copytree(misc_filepath, new_filepath, dirs_exist_ok=True)
 
         return ConfigurationPackage.from_toml(new_path_to_meta)
 
@@ -223,6 +240,7 @@ class ConfigurationPackage:
 
         for file_path in file_paths:
             try:
+                logging.info(f"Deleting {file_path}")
                 os.remove(file_path)
             except FileNotFoundError:
                 logging.error(f"Config file '{file_path}' not found.")
@@ -242,6 +260,7 @@ class ConfigurationPackage:
 
         for file_path in file_paths:
             try:
+                logging.info(f"Deleting {file_path}")
                 shutil.rmtree(file_path)
             except FileNotFoundError:
                 logging.error(f"Folder '{file_path}' not found.")
